make updates to location filter" "
