+++ app/(main)/events/page.tsx
@@ -9,7 +9,7 @@ import { Suspense } from "react";
 
 export const metadata: Metadata = {
   title: "Events",
-  description: "Popular Events Near New York",
+  description: "Popular Events",
 };
 
 export default function Page({
@@ -18,11 +18,11 @@ export default function Page({
   searchParams?: SearchParams;
 }) {
   const tag = searchParams?.tag || null;
-  const city = searchParams?.city || "new-york-ny";
+  const city = searchParams?.city || null;
   const distance = searchParams?.distance || 50;
 
-  let location = "New York, NY";
-  if (!cityMap[city]) {
+  let location = "United States";
+  if (city && !cityMap[city]) {
     const splitCity = city.split("-");
     const stateName = splitCity[splitCity.length - 1];
     const cityName = splitCity
@@ -30,7 +30,7 @@ export default function Page({
       .map((term) => capitalize(term))
       .join(" ");
     location = `${cityName}, ${stateName.toUpperCase()}`;
-  } else {
+  } else if (city) {
     location = cityMap[city].label;
   }
 
@@ -46,10 +46,12 @@ export default function Page({
         ) : (
           <h1 className="font-semibold text-2xl">Popular Events</h1>
         )}
-        <h3 className="font-semibold text-muted-foreground text-lg">
-          <span className="italic mr-1">Within {distance} miles of </span>
-          {location}
-        </h3>
+        {city && (
+          <h3 className="font-semibold text-muted-foreground text-lg">
+            <span className="italic mr-1">Within {distance} miles of </span>
+            {location}
+          </h3>
+        )}
       </div>
       <Events searchParams={searchParams} />
     </main>
